chore: refactor metadata tables into shared components
